pyproject.toml
[project]
name = "recommendation-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.7",
    "djangorestframework>=3.16.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "factory-boy>=3.3.3",
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "pytest-django>=4.11.0",
    "ruff>=0.11.2",
]
